backend/functions/pw_client/requirements.txt
functions-framework==3.*
google-cloud-secret-manager==2.*
google-cloud-firestore==2.*
requests==2.*
google-cloud-storage==2.*
